ai/dontcare/requirements.txt
absolufy-imports==0.3.1
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
alpaca-py==0.42.1
annotated-types==0.7.0
anyio==4.10.0
apscheduler==3.6.3
asttokens==3.0.0
attrs==25.3.0
authlib==1.6.3
beautifulsoup4==4.13.5
build==1.3.0
cachetools==4.2.2
ccxt==4.5.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.1
curl-cffi==0.13.0
cycler==0.12.1
cyclopts==3.24.0
cython==3.1.4
dateparser==1.2.2
decorator==5.2.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.116.1
fastmcp==2.12.3
fastuuid==0.12.0
filelock==3.19.1
fonttools==4.60.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
google-adk==1.13.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.113.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.37.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.57.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.36.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib-metadata==8.7.0
ipython==9.5.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyterlab-widgets==3.0.15
kiwisolver==1.4.9
lazy-object-proxy==1.12.0
litellm==1.77.1
llvmlite==0.44.0
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mcp==1.14.0
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.1
multidict==6.6.4
multitasking==0.0.12
mypy-extensions==1.1.0
narwhals==2.5.0
numba==0.61.2
numpy==2.2.6
openai==1.107.2
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pandas==2.3.2
pandas-ta==0.4.71b0
parse==1.20.2
parso==0.8.5
pathable==0.4.4
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
prompt-toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycares==4.11.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyparsing==3.2.5
pyperclip==1.9.0
pyproject-hooks==1.2.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==5.2.1
python-multipart==0.0.20
python-telegram-bot==13.15
pytz==2025.2
pyyaml==6.0.2
quantstats==0.0.77
ray==2.49.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.1
rsa==4.9.1
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.43
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
sse-starlette==3.0.2
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.47.3
ta==0.11.0
ta-lib==0.6.7
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
tornado==6.1
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vectorbt==0.28.1
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
werkzeug==3.1.1
widgetsnbextension==4.0.14
xmltodict==1.0.0
yarl==1.20.1
yfinance==0.2.66
zipp==3.23.0
